Add chatbot and notifications modules to feature integration

+++ homePresentation/src/commonMain/kotlin/com/spendoo/home/presentation/screen/HomeUiState.kt
@@ -131,5 +131,5 @@ fun CategorySpending.toUiState(): SpendingUiState {
 data class UserData(
     val userName: String,
     val userImageUrl: String?,
-    val notificationsCount: Int
+    val notificationsCount: Long
 )
